requirements.txt
streamlit
pdfminer.six
requests>=2.31.0
python-docx
nltk
spacy
sentence-transformers
scikit-learn
numpy
pandas
transformers
torch
faiss-cpu
openai
langchain
tiktoken
pydantic>=2.0
flask>=2.3.0
flask-cors>=4.0
python-dotenv>=1.0
pytest>=7.4
pytest-cov>=4.1
black>=23.0
flake8>=6.0
rank-bm25>=0.2.2
thefuzz>=0.19.0
python-Levenshtein>=0.21.0
